# app/docs/openapi.yaml
openapi: 3.1.0
info:
  title: whisperX REST service
  description: "\n    # whisperX RESTful API\n\n    Welcome to the whisperX RESTful\
    \ API! This API provides a suite of audio processing services to enhance and analyze\
    \ your audio content.\n\n    ## Documentation:\n\n    For detailed information\
    \ on request and response formats, consult the [WhisperX Documentation](https://github.com/m-bain/whisperX).\n\
    \n    ## Services:\n\n    Speech-2-Text provides a suite of audio processing services\
    \ to enhance and analyze your audio content. The following services are available:\n\
    \n    1. Transcribe: Transcribe an audio/video  file into text.\n    2. Align:\
    \ Align the transcript to the audio/video file.\n    3. Diarize: Diarize an audio/video\
    \ file into speakers.\n    4. Combine Transcript and Diarization: Combine the\
    \ transcript and diarization results.\n\n    ## Supported file extensions:\n \
    \   AUDIO_EXTENSIONS = {'.oga', '.ogg', '.mp3', '.wma', '.awb', '.amr', '.m4a',\
    \ '.wav', '.aac'}\n\n    VIDEO_EXTENSIONS = {'.mov', '.mp4', '.avi', '.mkv', '.wmv'}\n\
    \n    "
  version: 0.0.1
paths:
...
  /task/all:
    get:
      tags:
      - Tasks Management
      summary: Get All Tasks Status
      description: "Retrieve the status of all tasks.\n\nArgs:\n    service: Task\
        \ management service dependency.\n\nReturns:\n    TaskListResponse: The status\
        \ of all tasks."
      operationId: get_all_tasks_status_task_all_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/TaskStatusEnum'
          - type: 'null'
          description: Task status to filter by (queued, processing, completed, failed)
          title: Status
        description: Task status to filter by (queued, processing, completed, failed)
      - name: ref_user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ID of the user who created the task
          title: Ref User Id
        description: ID of the user who created the task
      - name: ref_request_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ID of the request associated with the task
          title: Ref Request Id
        description: ID of the request associated with the task
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /task/{identifier}:
    get:
      tags:
      - Tasks Management
      summary: Get Transcription Status
      description: "Retrieve the status of a specific task by its identifier.\n\n\
        Args:\n    identifier (str): The identifier of the task.\n    service: Task\
        \ management service dependency.\n\nReturns:\n    Result: The status of the\
        \ task.\n\nRaises:\n    TaskNotFoundError: If the identifier is not found."
      operationId: get_transcription_status_task__identifier__get
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /task/{identifier}/delete:
    delete:
      tags:
      - Tasks Management
      summary: Delete Task
      description: "Delete a specific task by its identifier.\n\nArgs:\n    identifier\
        \ (str): The identifier of the task.\n    service: Task management service\
        \ dependency.\n\nReturns:\n    Response: Confirmation message of task deletion.\n\
        \nRaises:\n    TaskNotFoundError: If the task is not found."
      operationId: delete_task_task__identifier__delete_delete
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /task/{identifier}/retry:
    post:
      tags:
      - Tasks Management
      summary: Retry Task
      description: "Retry a specific task by its identifier.\n\nArgs:\n    identifier\
        \ (str): The identifier of the task.\n    transcript (UploadFile): Uploaded\
        \ aligned transcript file.\n    diarization (UploadFile): Uploaded diarization\
        \ result file.\n    service: Task management service dependency.\n\nReturns:\n\
        \    Response: Confirmation message of task retrying.\n\nRaises:\n    TaskNotFoundError:\
        \ If the task is not found."
      operationId: retry_task_task__identifier__retry_post
      parameters:
      - name: identifier
        in: path
        required: true
        schema:
          type: string
          title: Identifier
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_retry_task_task__identifier__retry_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_retry_task_task__identifier__retry_post:
      properties:
        transcript:
          anyOf:
          - type: string
            format: binary
          - type: 'null'
          title: Transcript
        diarization:
          anyOf:
          - type: string
            format: binary
          - type: 'null'
          title: Diarization
      type: object
      title: Body_retry_task_task__identifier__retry_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Metadata:
      properties:
        task_type:
          type: string
          title: Task Type
        task_params:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Task Params
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
        duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration
        audio_duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Audio Duration
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
      type: object
      required:
      - task_type
      - task_params
      - language
      - file_name
      - url
      - duration
      title: Metadata
      description: Metadata model for task information.
    Response:
      properties:
        identifier:
          type: string
          title: Identifier
        message:
          type: string
          title: Message
      type: object
      required:
      - identifier
      - message
      title: Response
      description: Response model for API responses.
    Result:
      properties:
        status:
          type: string
          title: Status
        result:
          title: Result
        metadata:
          $ref: '#/components/schemas/Metadata'
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - status
      - result
      - metadata
      - error
      title: Result
      description: Model for a result with status, result data, metadata, and optional
        error.
...
    TaskListResponse:
      properties:
        tasks:
          items:
            $ref: '#/components/schemas/TaskSummaryResponse'
          type: array
          title: Tasks
          description: List of task summaries
      type: object
      required:
      - tasks
      title: TaskListResponse
      description: DTO for returning a list of task summaries.
    TaskStatusEnum:
      type: string
      enum:
      - queued
      - processing
      - completed
      - failed
      title: TaskStatusEnum
      description: Enum for task statuses.
    TaskSummaryResponse:
      properties:
        identifier:
          type: string
          title: Identifier
          description: Unique identifier for the task
        status:
          type: string
          title: Status
          description: Current status of the task
        task_type:
          type: string
          title: Task Type
          description: Type of task
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
          description: Name of the file
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
          description: URL of the file
        audio_duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Audio Duration
          description: Duration of the audio
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
          description: Language of the audio
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          description: Error message if task failed
        duration:
          anyOf:
          - type: number
          - type: 'null'
          title: Duration
          description: Execution duration in seconds
        start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Start Time
          description: Task start time
        end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: End Time
          description: Task end time
      type: object
      required:
      - identifier
      - status
      - task_type
      title: TaskSummaryResponse
      description: 'DTO for returning task summary in list operations.


        Used for GET /task/all endpoint to provide a lighter response.'
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...